server/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
python-dotenv
flask_cors
bcrypt
pymysql
